Early return if no ind sponsor role
Issue #976

+++ server/database/migrations/2016_05_25_130314_cleanup_old_independent_sponsor_data.php
@@ -20,6 +20,9 @@ class CleanupOldIndependentSponsorData extends Migration
         // Get ID for INDEPENDENT_SPONSOR role
         $role = DB::selectOne('select id from roles where name = ?', ['Independent Sponsor']);
 
+        // If there is no Independent Sponsor role go ahead and bail
+        if (!$roll) return;
+
         // Delete records in the `user_role` tabloe
         DB::delete('delete from role_user where role_id = ?', [$role->id]);
